+++ lib/cloudinaryConfig.js
@@ -1,7 +1,6 @@
 require('dotenv').config();
 const cloudinary = require('cloudinary').v2;
 
-
 cloudinary.config({
 	cloud_name: process.env.CLOUD_NAME,
 	api_key: process.env.CLOUD_API_KEY,
