+++ route/movieRouter.js
@@ -1,18 +1,16 @@
 const express = require('express');
-const router = express.Router();
-const movieController = require('../controller/MovieController');
 const multer = require('multer');
-// const { videoUpload } = require('../config/multerConfig');
+const router = express.Router();
+const MovieController = require('../controller/MovieController');
 
 const fileUpload = multer();
+const movie = new MovieController();
 
-const movie = new movieController();
+router.get('/movies', movie.getMovies);
 
-router.get('/movies', movie.getMovie);
+router.get('/movies/:id', movie.getMovieById);
 
-router.get('/movies/:id', movie.searchMovieById);
-
-router.post('/movies/upload',fileUpload.single('video'), movie.createMovie);
+router.post('/movies/upload', fileUpload.single('video'), movie.createMovie);
 
 router.put('/movies/:id', movie.updateMovie);
